+++ src/Application/Candidates/Commands/Update/UpdateCandidateCommand.cs
@@ -14,7 +14,7 @@ namespace Application.Candidates.Commands.Update
         
         public string FirstName { get; set; }
 
-        public int DateOfBirth { get; set; }
+        public long DateOfBirth { get; set; }
 
         public string Surname { get; set; }
